Update and rename br_lem_ba.py to lem_br_ba.py
